tests: add proxy option to test file for driver
Adding a proxy option to the driver/test_kubeflow_workloads.py script, will allow users to tell
the driver it will run in a proxied environment. With this, a PodDefault will be used for setting
environment variables (NO_PROXY, HTTP_PROXY, HTTPS_PROXY) in the Pod(s) that execute the Job for UATs.

Fixes #96